Add files via upload
Attempts to model login via signing key verification. This is password-based and may be a similar approach to Metamask.

A new user is issued a public key and a private key.

The public key is stored in a DB.
The private key is encrypted and stored locally on the user's computer.

A successful log-in is decrypting the signing key and verifying that it can produce the public key. All using the Cardano-cli.